+++ crates/jsonschema/src/keywords/unevaluated_properties.rs
@@ -4,7 +4,7 @@ use crate::{
     keywords::CompilationResult,
     node::SchemaNode,
     output::BasicOutput,
-    paths::{JsonPointerNode, Location},
+    paths::{LazyLocation, Location},
     primitive_type::PrimitiveType,
     properties::*,
     validator::{PartialApplication, Validate},
@@ -187,8 +187,8 @@ impl UnevaluatedPropertiesValidator {
     fn validate_property<'instance>(
         &self,
         instance: &'instance Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -262,8 +262,8 @@ impl UnevaluatedPropertiesValidator {
     fn apply_property<'a>(
         &'a self,
         instance: &Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &Value,
         property_name: &str,
     ) -> Option<BasicOutput<'a>> {
@@ -350,7 +350,7 @@ impl Validate for UnevaluatedPropertiesValidator {
     fn validate<'instance>(
         &self,
         instance: &'instance Value,
-        instance_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
     ) -> ErrorIterator<'instance> {
         if let Value::Object(props) = instance {
             let mut errors = vec![];
@@ -392,7 +392,7 @@ impl Validate for UnevaluatedPropertiesValidator {
     fn apply<'a>(
         &'a self,
         instance: &Value,
-        instance_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
     ) -> PartialApplication<'a> {
         if let Value::Object(props) = instance {
             let mut output = BasicOutput::default();
@@ -461,7 +461,7 @@ impl PropertySubvalidator {
 
     fn validate_property<'instance>(
         &self,
-        property_path: &JsonPointerNode,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -472,7 +472,7 @@ impl PropertySubvalidator {
 
     fn apply_property<'a>(
         &'a self,
-        property_path: &JsonPointerNode,
+        property_path: &LazyLocation,
         property_instance: &Value,
         property_name: &str,
     ) -> Option<BasicOutput<'a>> {
@@ -518,7 +518,7 @@ impl PatternSubvalidator {
 
     fn validate_property<'instance>(
         &self,
-        property_path: &JsonPointerNode,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -538,7 +538,7 @@ impl PatternSubvalidator {
 
     fn apply_property<'a>(
         &'a self,
-        property_path: &JsonPointerNode,
+        property_path: &LazyLocation,
         property_instance: &Value,
         property_name: &str,
     ) -> Option<BasicOutput<'a>> {
@@ -684,8 +684,8 @@ impl SubschemaSubvalidator {
     fn validate_property<'instance>(
         &self,
         instance: &'instance Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -766,8 +766,8 @@ impl SubschemaSubvalidator {
     fn apply_property<'a>(
         &'a self,
         instance: &Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &Value,
         property_name: &str,
     ) -> Option<BasicOutput<'a>> {
@@ -885,7 +885,7 @@ impl UnevaluatedSubvalidator {
 
     fn validate_property<'instance>(
         &self,
-        property_path: &JsonPointerNode,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         _property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -900,7 +900,7 @@ impl UnevaluatedSubvalidator {
 
     fn apply_property<'a>(
         &'a self,
-        property_path: &JsonPointerNode,
+        property_path: &LazyLocation,
         property_instance: &Value,
         _property_name: &str,
     ) -> Option<BasicOutput<'a>> {
@@ -1006,8 +1006,8 @@ impl ConditionalSubvalidator {
     fn validate_property<'instance>(
         &self,
         instance: &'instance Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -1050,8 +1050,8 @@ impl ConditionalSubvalidator {
     fn apply_property<'a>(
         &'a self,
         instance: &Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &Value,
         property_name: &str,
     ) -> Option<BasicOutput<'a>> {
@@ -1148,8 +1148,8 @@ impl DependentSchemaSubvalidator {
     fn validate_property<'instance>(
         &self,
         instance: &'instance Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -1173,8 +1173,8 @@ impl DependentSchemaSubvalidator {
     fn apply_property<'a>(
         &'a self,
         instance: &Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &Value,
         property_name: &str,
     ) -> Option<BasicOutput<'a>> {
@@ -1258,8 +1258,8 @@ impl ReferenceSubvalidator {
     fn validate_property<'instance>(
         &self,
         instance: &'instance Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &'instance Value,
         property_name: &str,
     ) -> Option<ErrorIterator<'instance>> {
@@ -1275,8 +1275,8 @@ impl ReferenceSubvalidator {
     fn apply_property<'a>(
         &'a self,
         instance: &Value,
-        instance_path: &JsonPointerNode,
-        property_path: &JsonPointerNode,
+        instance_path: &LazyLocation,
+        property_path: &LazyLocation,
         property_instance: &Value,
         property_name: &str,
     ) -> Option<BasicOutput<'a>> {